feat: remove legacy storage
