+++ test/e2e/api_transactions_list.go
@@ -333,7 +333,7 @@ var _ = Context("Ledger transactions list API tests", func() {
 				Expect(err).To(HaveOccurred())
 			})
 			It("Should fail with "+string(components.V2ErrorsEnumInternal)+" error code", func() {
-				Expect(err).To(HaveErrorCode(string(components.V2ErrorsEnumInternal)))
+				Expect(err).To(HaveErrorCode(string(components.V2ErrorsEnumValidation)))
 			})
 		})
 	})
